refactor: move tags to templatetags folder
