+++ templates/forum/thread.haml
@@ -1,6 +1,7 @@
 -extends "base.haml"
 -load static
 -load compress
+-load forum_tags
 
 -block head
   -compress css
